+++ coordinator/src/test/scala/filodb.coordinator/queryplanner/PlannerHierarchySpec.scala
@@ -1961,13 +1961,6 @@ class PlannerHierarchySpec extends AnyFunSpec with Matchers with PlanValidationS
       }
     }
 
-    //  Sum()
-    //    Stitch()
-     //     RemoteExec(test{job="app"}[])
-     //     RemoteExec(test{job="app"}[])
-    //
-
-
     val tests = Seq(
       // aggregate
       Test("""sum(test{job="app"} offset 10m)""",
